requirements.txt
about-time==4.2.1
alive-progress==3.1.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.1.0
autograd==1.5
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cma==3.2.2
cmake==3.26.3
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==40.0.2
cvxopt==1.3.1
cvxpy==1.3.1
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.6
docker-pycreds==0.4.0
ecos==2.0.12
exceptiongroup==1.1.1
filelock==3.12.0
fonttools==4.39.4
future==0.18.3
geotorch==0.3.0
gitdb==4.0.10
GitPython==3.1.31
grapheme==0.6.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.5
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
meshzoo==0.11.1
mpmath==1.3.0
networkx==3.1
ntplib==0.4.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
osqp==0.6.2.post9
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
platformdirs==3.5.1
protobuf==4.23.1
psutil==5.9.5
py-machineid==0.3.0
pycparser==2.21
Pygments==2.15.1
pymoo==0.6.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-license-manager==0.2.6
pytz==2023.3
PyYAML==6.0
qdldl==0.1.7
requests==2.31.0
requests-cache==1.0.1
rich==13.3.5
rich_argparse==1.1.0
rotograd==0.1.5.2
scikit-learn==1.2.2
scipy==1.10.1
scs==3.2.3
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sympy==1.12
threadpoolctl==3.1.0
tomli==2.0.1
tomli_w==1.0.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.10.3
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
tzdata==2023.3
url-normalize==1.4.3
urllib3==1.26.16
wandb==0.15.3
winregistry==1.1.1
wrapt==1.15.0
x21==0.4.12
